+++ package.json
@@ -15,6 +15,8 @@
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-icons": "^5.5.0",
+    "react-router": "^7.3.0",
+    "react-use": "^17.6.0",
     "shogi.js": "^5.4.1"
   },
   "devDependencies": {
