--- src/App.css
@@ -1,157 +1,0 @@
-#root {
-	max-width: 1280px;
-	margin: 0 auto;
-	padding: 2rem;
-	text-align: center;
-}
-
-.logo {
-	height: 6em;
-	padding: 1.5em;
-	will-change: filter;
-	transition: filter 300ms;
-}
-
-.logo:hover {
-	filter: drop-shadow(0 0 2em #646cffaa);
-}
-
-.logo.react:hover {
-	filter: drop-shadow(0 0 2em #61dafbaa);
-}
-
-@keyframes logo-spin {
-	from {
-		transform: rotate(0deg);
-	}
-	to {
-		transform: rotate(360deg);
-	}
-}
-
-@media (prefers-reduced-motion: no-preference) {
-	a:nth-of-type(2) .logo {
-		animation: logo-spin infinite 20s linear;
-	}
-}
-
-.card {
-	padding: 2em;
-}
-
-.read-the-docs {
-	color: #888;
-}
-.style {
-	display: flex;
-	flex-direction: column;
-	justify-content: center;
-	align-items: center;
-	height: 100vh;
-	/* width: 100vw; */
-	position: relative;
-}
-
-.cite_style {
-	width: 100vw;
-	height: 90vh;
-	display: flex;
-	flex-direction: column;
-	justify-content: center;
-	align-items: center;
-	margin: 0;
-}
-
-.container {
-	display: flex;
-	flex-direction: column;
-	align-items: left;
-	margin-left: 20px;
-}
-.container-bram {
-	display: flex;
-	flex-direction: column;
-	align-items: right;
-	margin-right: 5px;
-}
-.hand-wrapper-bram {
-	width: 300px;
-	height: 40px;
-	display: flex;
-	flex-wrap: wrap;
-	gap: 5px;
-	justify-content: flex-end;
-}
-.hand-wrapper {
-	width: 300px;
-	height: 40px;
-	display: flex;
-	flex-wrap: wrap;
-	gap: 5px;
-}
-
-.hand-piece {
-	width: 40px;
-	height: 40px;
-	color: black;
-	display: flex;
-	align-items: center;
-	justify-content: flex-start;
-}
-.hand-piece-bram {
-	width: 40px;
-	height: 40px;
-	color: black;
-	display: flex;
-	align-items: center;
-	justify-content: flex-end;
-	transform: rotate(180deg); /* 持ち駒を反転表示 */
-}
-
-.hand-piece.enabled {
-	cursor: pointer;
-	pointer-events: auto;
-}
-
-.hand-piece.disabled {
-	cursor: not-allowed;
-	pointer-events: none;
-}
-
-.hand-piece.selected {
-	background-color: yellow;
-}
-
-.board {
-	display: grid;
-	grid-template-rows: repeat(9, 50px);
-	grid-template-columns: repeat(9, 50px);
-	gap: 0.5px;
-	justify-items: center;
-	align-items: center;
-}
-
-.cell {
-	width: 50px;
-	height: 50px;
-	display: flex;
-	align-items: center;
-	justify-content: center;
-	font-size: 20px;
-	color: black;
-	border: 1px solid black;
-	cursor: pointer;
-	background-color: gray;
-}
-
-.cell.highlighted {
-	background-color: lightblue;
-}
-
-.cell.selected {
-	background-color: yellow;
-}
-
-.cell.rotated {
-	transform: rotate(180deg);
-}
