+++ src/pages/Home.module.css
@@ -1,0 +1,66 @@
+.container {
+	width: 100dvw;
+	height: 100%;
+	display: flex;
+	flex-direction: column;
+	gap: 1em;
+	align-items: center;
+	justify-content: center;
+	background-color: #abc49a;
+}
+
+.title {
+	font-size: 4em;
+}
+
+.button {
+	--color: #f0f0f0;
+	--color-hover: #df6722;
+
+	position: relative;
+	width: 400px;
+	height: 48px;
+	border: none;
+	border-radius: 8px;
+	font-size: 1.2em;
+	background-color: var(--color);
+	box-shadow: 0 0 16px 0 rgba(0, 0, 0, 0.4);
+	overflow: hidden;
+	transition: 1s;
+
+	&:hover {
+		&:after {
+			height: 100%;
+		}
+		&:before {
+			height: 100%;
+		}
+	}
+
+	&::before {
+		content: "";
+
+		position: absolute;
+		left: 0;
+		top: 0;
+		width: 100%;
+		height: 0;
+		background-color: var(--color-hover);
+		mix-blend-mode: color-burn;
+
+		transition: 1s;
+	}
+
+	&::after {
+		content: "";
+		position: absolute;
+		left: 0;
+		bottom: 0;
+		width: 100%;
+		height: 0;
+		background-color: var(--color-hover);
+		mix-blend-mode: color-burn;
+
+		transition: 1s;
+	}
+}
